+++ examples/basic/controllers/home_controller.rb
@@ -1,7 +1,0 @@
-require '../../lib/radical'
-
-class HomeController < Controller
-  def index
-    render plain: 'home#index!'
-  end
-end