+++ test/app.rb
@@ -1,14 +1,11 @@
 require 'radical'
 
-Dir[File.join(__dir__, 'controllers', '*_controller.rb')].each do |file|
-  require file
-end
+require_relative 'controllers/home'
+require_relative 'controllers/todos'
+require_relative 'controllers/todo_items'
 
-class App < Radical
-  routes do
-    get '/', to: 'home#index'
-    get '/todos', to: 'todos#index'
-    get '/todos/:id/edit', to: 'todos#edit'
-    get '/todo-items/new', to: 'todo_items#new'
-  end
+class App < Radical::App
+  routes Home
+  routes Todos
+  routes TodoItems
 end