+++ test/controllers/home.rb
@@ -1,0 +1,7 @@
+require 'radical'
+
+class Home < Radical::Controller
+  def index
+    @page = 'home#index'
+  end
+end