+++ test/controllers/home_controller.rb
@@ -1,4 +1,0 @@
-require 'radical'
-
-class HomeController < Controller
-end