+++ test/controllers/todo_items.rb
@@ -1,0 +1,7 @@
+require 'radical'
+
+class TodoItems < Radical::Controller
+  def new
+    plain 'todo_items#new'
+  end
+end