+++ test/controllers/todo_items_controller.rb
@@ -1,7 +1,0 @@
-require 'radical'
-
-class TodoItemsController < Controller
-  def new
-    render plain: 'todo_items_controller'
-  end
-end