+++ test/test_radical.rb
@@ -1,6 +1,9 @@
 require 'minitest/autorun'
 require 'app'
 
+class NoRoutesApp < Radical::App
+end
+
 class RadicalTest < Minitest::Test
   def setup
     @app = App.new
@@ -9,14 +12,16 @@ class RadicalTest < Minitest::Test
   def teardown; end
 
   def test_app_returns_a_404_with_no_routes
-    assert_equal [404, { 'Content-Type' => 'text/plain' }, ['404 Not Found']], @app.call({})
+    no_routes_app = NoRoutesApp.new
+
+    assert_equal [404, {}, []], no_routes_app.call({})
   end
 
-  def test_returns_expected_status_code_from_route
-    expected = [404, { 'Content-Type' => 'text/plain' }, ['404 Not Found']]
+  def test_returns_404_when_route_not_found
+    expected = [404, {}, ['404 Not Found']]
     actual = @app.call(
       {
-        'PATH_INFO' => '/',
+        'PATH_INFO' => '/404',
         'REQUEST_METHOD' => 'GET'
       }
     )
@@ -24,8 +29,8 @@ class RadicalTest < Minitest::Test
     assert_equal expected, actual
   end
 
-  def test_returns_diff_route
-    expected = [200, { 'Content-Type' => 'text/plain' }, ['todos']]
+  def test_returns_200_when_route_is_found
+    expected = [200, { 'Content-Type' => 'text/plain' }, ['todos#index']]
     actual = @app.call(
       {
         'PATH_INFO' => '/todos',
@@ -36,11 +41,11 @@ class RadicalTest < Minitest::Test
     assert_equal expected, actual
   end
 
-  def test_routes_to_long_controller
-    expected = [200, { 'Content-Type' => 'text/plain' }, ['todo_items_controller']]
+  def test_routes_to_long_controller_name
+    expected = [200, { 'Content-Type' => 'text/plain' }, ['todo_items#new']]
     actual = @app.call(
       {
-        'PATH_INFO' => '/todo-items/new',
+        'PATH_INFO' => '/todo_items/new',
         'REQUEST_METHOD' => 'GET'
       }
     )
@@ -49,7 +54,7 @@ class RadicalTest < Minitest::Test
   end
 
   def test_params
-    expected = [200, { 'Content-Type' => 'text/plain' }, ['todos#edit id: 2']]
+    expected = [200, { 'Content-Type' => 'text/plain' }, ['todos#edit { id: 2 }']]
     actual = @app.call(
       {
         'PATH_INFO' => '/todos/2/edit',